DDP-8192 unite brugada consent patches
