--- pepper-apis/studybuilder-cli/studies/brugada/patches/text-revisions.conf
@@ -1,15 +1,0 @@
-"study":{
-  "guid":"brugada"
-}
-"translation-updates":[
-  {
-    "activityCode": "CONSENT"
-    "varName" : "consent_research_p6",
-    "newValue" : "If you have any questions before you sign this consent form or after you join the study, you can contact the tissue bank staff at <a class=\"underlined-link\" href=\"tel:6177147560\">617-714-7560</a> from Monday - Friday 9 am – 5pm. The person in charge of the tissue bank is Patrick Ellinor, MD, PhD.",
-  },
-  {
-    "activityCode": "MEDICAL_RECORD_RELEASE"
-    "varName" : "medical_record_release_s3li7",
-    "newValue" : "I have already read and signed the informed consent document for this study, which describes the use of my personal health information, and hereby grant permission to Patrick Ellinor, MD, PhD, and/or members of the study team to examine copies of my medical records pertaining to my Brugada syndrome diagnosis and other related conditions. I acknowledge that I can access a copy of this completed form on my dashboard on the study website at brugada.linkmyheart.org."
-  },
-]
